update dependencies when including/excluding a crate
This will update the workspace dependencies/dev-dependencies/build-dependencies
of a crate on exclude and include commands.

On exclude, all workspace dependencies are inlined in the crate's Cargo.toml so
that it can be build standalone.

On include, dependencies that are compatiple with the workspace.dependencies are
made to use those.